[change] Separated the serializer classes from views #414
- Moved the serializer classes from views.py file to a separate serializers.py file.

closes #414
